YAN-1966 Allow specifying a securityContext
While in principle it shouldn't occur, it is often the case that a PVC
has data with permissions that makes it impossible to be read by
non-root users.

This commit introduces a escape hatch for this issue, with Helm values
being able to override the pod's securityContext. This in turn allows
users installing the chart to specify that the containers should run as
a certain user (e.g., root), or the GID used to use for filesystem
operations.

Signed-off-by: Rodrigo Tobar <rtobar@icrar.org>